+++ CopilotKit/packages/react-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # ui
 
+## 0.13.0
+
+### Minor Changes
+
+- contextCategories no longer optional for reading context
+
+## 0.12.0
+
+### Minor Changes
+
+- fixed bug: useMakeCopilotDocumentReadable category reference
+
 ## 0.11.0
 
 ### Minor Changes