+++ CopilotKit/packages/react-textarea/CHANGELOG.md
@@ -1,5 +1,27 @@
 # ui
 
+## 0.24.0
+
+### Minor Changes
+
+- contextCategories no longer optional for reading context
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-core@0.13.0
+
+## 0.23.0
+
+### Minor Changes
+
+- fixed bug: useMakeCopilotDocumentReadable category reference
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-core@0.12.0
+
 ## 0.22.0
 
 ### Minor Changes